+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Kit Editor: bank switching buttons.
 
+### Fixed
+ - Error handling when palettes cannot be parsed.
+
 ## [1.10.2] - 2020-11-22
 ### Fixed
  - Kit Editor: when replacing samples, trim sample end to fit.